Optimized WebClient calls
Fixed WebClient call  w/o a dispose, and changed the using pattern into try catch finally dispose (due to using not actually freeing up the underlying socket for HttpClient).

WebClient is a wrapper around HttpWebRequest.  But I'm not sure if the problem of HttpClient extends back to that.  I changed it nonetheless to be sure.  I remember a bunch of networking glut in the Cinema APIs which used it extensively, and I'm not sure if they are all disposed of properly if at all.

Refer to:

https://docs.microsoft.com/en-us/dotnet/architecture/microservices/implement-resilient-applications/use-httpclientfactory-to-implement-resilient-http-requests

https://aspnetmonsters.com/2016/08/2016-08-27-httpclientwrong/

https://www.infoworld.com/article/3198673/when-to-use-webclient-vs-httpclient-vs-httpwebrequest.html#:~:text=In%20a%20nutshell%2C%20WebRequest%E2%80%94in,and%20posts%2C%20having%20arrived%20with%20.